waypoint ckin

+++ test_opengl/shaders/glut_based/tut06_textures_ex/src/Program.cpp
@@ -49,7 +49,7 @@ VarAttribute* Program::get_var_attribute(const GLchar *name) const
 VarUniform* Program::get_var_uniform(const GLchar *name) const
 {
     VarUniform* var_uniform = new VarUniform(this, name);
-    if(var_uniform && var_uniform->id() == -1)
+    if(var_uniform && var_uniform->id() == static_cast<GLuint>(-1))
     {
         fprintf(stderr, "Could not bind uniform %s\n", name);
         delete var_uniform;
